API Type Checking
This adds editor-supported type checking which validates how the code is written, corresponding to the types from the APIs on npm. This makes it easier to write scripts around the APIs because they will autocomplete in your editor when using them, and it will also announce to you when it isn't being used quite as expected. Super awesome!

Note that this is still JavaScript, yet it can behave like TypeScript. So functionally this is usable just like TypeScript, but you don't need to compile your code to enable this type checking. It simply shows in your editor instead, no custom syntax needed.

The changes in this commit are carried over from a project that I started working on a few months ago, which looks into similar experimentation.

https://github.com/Offroaders123/Bedrock-Script-API/tree/dba201420db2a8c30987a4e7069721fcf1350afc