Form Value Usage Assertion
Added a type assertion to make the `formValues` usage more accurate to the values that it is using at runtime. This is a common thing to do when type-checking is enabled, sometimes we as the developers know more than the compiler is able to know.

It is safe to assert the type here, because we know how the code above it is implemented, and what those values will be. So we have more information than it can infer for us, so we can let it know of that information so it can do things according to that.

Forgot to include this link in the previous commit, where I fixed the beta std library typings.

https://stackoverflow.com/questions/78213780/minecraft-bedrock-scripting-api-error-world-beforeevents-chatsend-subscribe